ambient-light/AmbientLightSensor_onerror-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AmbientLightSensor Test: onerror</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/ambient-light/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<h2>Precondition</h2>
<ol>
  <li>
    Disable the Ambient Light Sensor or run test on a device without Amibent Light Sensor.
  </li>
</ol>
<div id="log"></div>
<script>
let sensor1, sensor2;
setup(function() {
  sensor1 = new AmbientLightSensor();
  sensor2 = new AmbientLightSensor();
});

async_test(function(t) {
  sensor1.onstatechange = t.step_func_done(function(event) {
    if ( sensor1.state == "errored" ) {
      assert_true(true);
    }
  });
  sensor1.start();
}, "Test that sensor state changes to 'errored' when sensor is not supported");

async_test(function(t) {
  sensor2.onchange = t.step_func_done(function(event) {
    assert_unreached();
  });
  sensor2.onerror = t.step_func_done(function(event) {
    assert_equals(sensor2.state, 'errored');
    assert_equals(event.error.name, 'NotFoundError');
  });
  sensor2.start();
}, "Test that 'onerror' is send when sensor is not supported");

</script>
